Split Product.js into multiple files, for each component
