+++ products/products.css
@@ -398,20 +398,11 @@
 .product-part__link {
   position: relative;
   width: fit-content;
+  text-decoration: underline;
+  text-decoration-thickness: 0.2rem;
   letter-spacing: 0.08rem;
   font-weight: 900;
   font-size: 1.6rem;
-}
-
-.product-part__link::after {
-  content: "";
-  display: inline-block;
-  position: absolute;
-  bottom: 0;
-  left: 0;
-  width: 100%;
-  height: 0.2rem;
-  background-color: var(--outer-space);
 }
 
 .reviews {
