+++ products/tree-runners.html
@@ -358,6 +358,6 @@
         </p>
       </div>
     </section>
-    <script type="module" src="../scripts/Product.js"></script>
+    <script type="module" src="../scripts/Product/index.js"></script>
   </body>
 </html>
