+++ scripts/Product/index.js
@@ -1,0 +1,6 @@
+import "./Product.js";
+import "./Carousel.js";
+import "./ColorPicker.js";
+import "./SizePicker.js";
+import "./Collapsible.js";
+import "./Reviews.js";
